refactor(sound): collapse shared Web Audio service
