+++ apps/whispering/src/lib/services/index.ts
@@ -1,10 +1,10 @@
 import { AnalyticsServiceLive } from '#platform/analytics';
 import { AudioBlobStoreLive } from '#platform/blob-store';
 import { DownloadServiceLive } from '#platform/download';
-import { PlaySoundServiceLive } from '#platform/sound';
 import { TextServiceLive } from '#platform/text';
 import * as completions from './completion';
 import { LocalShortcutManagerLive } from './local-shortcut-manager';
+import { PlaySoundServiceLive } from './sound';
 
 /**
  * Cross-platform services.
