--- apps/whispering/src/lib/services/sound/index.browser.ts
@@ -1,9 +1,0 @@
-import { soundSources } from './assets';
-import { createWebAudioPlaySoundService } from './web-audio';
-
-export type { PlaySoundService, SoundError } from './types';
-
-export const PlaySoundServiceLive = createWebAudioPlaySoundService({
-	shouldPlay: () => !document.hidden,
-	soundSources,
-});
