--- apps/whispering/src/lib/services/sound/index.tauri.ts
@@ -1,8 +1,0 @@
-import { soundSources } from './assets';
-import { createWebAudioPlaySoundService } from './web-audio';
-
-export type { PlaySoundService, SoundError } from './types';
-
-export const PlaySoundServiceLive = createWebAudioPlaySoundService({
-	soundSources,
-});
